requirements.txt
pyyaml~=6.0
loguru~=0.6.0